Raising TaxonNotFoundException in GraphBase.findTaxNodeByName and changes associated with MainRunner errors.
My apologies for hitting so many lines with this commit...
As I prepare for the REST webservices to call many of the functions used by MainRunner,
it is important for System.exit() to not be called in these functions.

The early exits without finally clauses which call shutdownDB was causing my
db to be not shutdown properly in many cases (I'm afraid that I often have syntax
errors in my invocation of treemachine).

Some exceptions were being suppressed, and treemachine was exiting with rc 0 in
some cases in which the command failed. This makes it hard to chain commands together
in bash (using bash's if to test for success or the && operator in bash).

So, I've refactored the commands in MainRunner to return 0 if they succeed. If they
return 1, then there was an error. 2 means that there was an error and the calling
function should show the general help message.

In functions that appear to use the db, I've put the shutdownDB in a finally block
so that it will be called even if there is an exception.

This commit was triggered by me trying to figure out why:
    ./treemachine addtree ex.nexson Rosids a test.db
gave me a NullPointerException.  It turns out that I should have been using:
    ./treemachine addtree ex.nexson rosids a test.db