Storing tree id's from phylografter.
treeID is obtained from nexml.trees.tree.@id in NEXSON. The prefix "tree" (if it is found) is trimmed from this string; this makes the ID agree with the phtylografter ID for the tree.

When reading a newick file treeID must be obtained from another file.

In the MainRunner: addtree command has been split into addnewick and addnexson (with addnewick taking an additional argument for the file of ids).  These functions aren't tested yet.

The SourceRootNodesIndex now will have the root of each sourcetree associated with both rootnode -> sourcename (old behavior) and rootnodeforid -> treeID mappings.

The metadatanode now has a treeID property.

In a fully specified JadeTree object, the id is stored using assocObject("id", treeID)