Reverted back to older style of traversing a source tree.
This will entail grabbing all of the relationships in a tree even if
the caller only requests a thin slice of the tree. However, I was
having trouble with the using the contained taxon ids in original_taxa_map
not being found in the mrca of the GoL nodes. The FilterByPropertyRelIterator.java
file is currently not used (it has to be there to compile, but it is called
in a if (false) branch of code). It can be removed if we're confident that
it won't be useful.