Fix to ```getDraftTreeMRCAForNodes```.
When a query taxon is not present in the synthetic tree (i.e. it is non-monophyletic), it does not lie on a synthesis path,
and so cannot contribute to synth-MRCA calculations. Rather than ignore such problems, the updated method find the taxonomic
tip descendants of the non-monophyletic taxon, and uses them in the MRCA calculation.

Rationale (example): if a query is for the MRCA of 2 mammal taxa and a non-monophyletic avian taxon, if we drop the avian taxon
from consideration, the calculated MRCA will be within mammals, which is not the desired result. By using the taxonomic tip
descendants of the avian taxon in the calculation, the desired MRCA will be returned.

This is a nice complement to another service ```getTaxonomyMRCAForNodes```, which returns the taxonomic MRCA. A comparison of
the two calculated MRCAs could be very useful, e.g. in showing how poor the taxonomy is.