renamed STED support to ITED
"source tree" -> "input tree" just to cut down on jargon that is
commonly used within the open tree of life project in favor of
more widely understood phrases. Because it is common to refer to the
source and destination of edges, the term "source tree" can be a bit
confusing when talking about merging trees into graphs.