Moved synth test harness over from test-refactor
That branch was the descendant of a branch on which
I had made some synthesis fixes to deal with the early
failing tests.

When I look at a pull request between test-refactor and
the rootward-synth or issue-164 branch, I see lots of
changes that could affect synthesis. I'm assuming that the
were probably intentionally not included by CH and SAS on
this branch.

I should be cherry-picking the commits that just involve
tests, but I'm not sure that I was conscientious about
just making changes that affect the test harness.

So, I just copied the new test content over to this branch - which
will lead to some "this file is already there" warnings/errors
if we do merge those branches.