Another test. Fails for MTH - not sure if I'm running this correctly.
Inputs:
        ((A,B),E)
        (((A,B),C),E)
        (((A,B),D),E)
Expected output:
        (((A,B),C,D),E)
Obtained output:
        ((((A,B),D),C),E)

I don't really understand all of the steps of the current synth, but
I think that the issue here is that the creation of the set of
statements that span multiple trees (denoted S in the google doc),
only the root of one tree is combined with the internal nodes of the
other trees. Based on my limited understanding of the TAG, I did not
understand how a node would be created with the ingroup of ABCD. That
node would be created by merging the (non-root) internals of 2nd and
3rd trees. So, that was the motivation for designing this test.

I did not really have an expectation of what tree would be returned,
but there are 3 trees that are compatible with all of the input splits.
The expected tree is the strict consensus of them.

No input supports the statement that D is closer to A+B than C is, so
I don't think that the current behavior is correct.